refactor: Rename search screen to list. It will be used for wishlist too.
