+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".presentation.search.SearchActivity">
+        <activity android:name=".presentation.list.ListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
