+++ app/src/main/java/de/mytoysgroup/movies/challenge/presentation/list/ListAdapter.kt
@@ -1,4 +1,4 @@
-package de.mytoysgroup.movies.challenge.presentation.search
+package de.mytoysgroup.movies.challenge.presentation.list
 
 import android.support.annotation.LayoutRes
 import android.support.v7.widget.RecyclerView
